+++ eps_module.f90
@@ -2,5 +2,5 @@
 
 module eps_module
   implicit none
-  real, dimension(:), allocatable :: ex, ey, ez, epx, epy, epz, ebmag, epmag, ekg
+  double precision, dimension(:), allocatable :: ex, ey, ez, epx, epy, epz, ebmag, epmag, ekg
 end module eps_module
