+++ lcfs_module.f90
@@ -5,7 +5,7 @@ module lcfs_module
   integer :: num_lcfs=0
   integer, dimension(2) :: lcfs_size
 
-  real, allocatable, dimension(:, :, :) :: lcfs
+  double precision, allocatable, dimension(:, :, :) :: lcfs
   character*144 :: lcfs_file=''
 
 end module lcfs_module
