+++ options_module.f90
@@ -7,9 +7,9 @@ module options_module
   integer :: use_diffusion=0
   integer :: diffusion_species=0
   integer :: output_coils=0, varD=0
-  real :: dpar1, dpar2
-  real :: d_perp, temperature, boozer_step, boozer_phi
-  real :: varDgamma=1.0, varDB0=1.0
+  double precision :: dpar1, dpar2
+  double precision :: d_perp, temperature, boozer_step, boozer_phi
+  double precision :: varDgamma=1.0, varDB0=1.0
 
 
 end module options_module
